refactor(permission): inline absolute-form helper and tighten path tests
Fold the single-use absoluteForm() helper into evaluatePattern() per the
repo convention, and correct the comment to only claim what the code does
(URLs are excluded, not bare command strings).

Rework the unit tests: replace a case that duplicated the fail-closed test
with one pinning that relative rules stay worktree-scoped, and add a
baseline assertion showing evaluate() still misses the absolute deny rule
that evaluatePattern() now catches. Move the file under test/permission/.
